retry: only buffer request bodies when retries are enabled (#1020)
PR #1017 added support for retries on requests with bodies. This
requires buffering the body data so that it can be sent as part of a
retry request.

In #1017, all requests in the profile HTTP route stack have their bodies
wrapped in a `ReplayBody` type that buffers the request data. Although
the buffering strategy is lazy (the request can be sent before the
client finishes sending the body, so we don't introduce the latency of
waiting to buffer the entire body) and fairly lightweight (the buffered
body data is not copied an additional time, and is cloned only by
increasing the reference count on the buffer it was read into from the
wire), there is still some additional overhead that could be avoided
when retries are not configured on a route.

This branch updates the `Retry` service so that buffering is only added
on routes which have a retry policy. If there is no retry policy, we
avoid wrapping the request's body in a `ReplayBody`.

The `retry` crate is now called `http-retry`.