inbound: Simplify protocol-detection skipping
The inbound proxy's layer that switches between routed and passthru
stacks uses an unnecessary explict type for its predicate. In order to
setup for further changes to this stack, this change eliminate the
specialized type in favor of a closure.

We also ditch the use of IndexSet as there's no functional need for it.