Implement reconnect as a NewService
In order to make the reconnect module more flexible--so that it can more
easily be used in different parts of the stack--this change modifies the
`reconnect` crate to be implemented in terms of `NewService` rather than
a `MakeService`.

Furthermore, tests are added to exercise reconnect.

This change sets up changes to reconnect behavior so reconnection is
only used when service discovery is responsible for evicting endpoints
from a balancer.