service-profiles: Wrap receiver types
We should be able to cache profile resolutions independently of the
stack; but the fact that the the profile modue's `Receiver` type is just
an alias for `watch::Receiver<Profile>` means that we can't attach
the additional metadata needed for cache entry retention.

In anticipation of caching profiles by lookup address, this change
creates an explicit `Receiver` type that hides the underlying
`watch::Receiver`, exposing only methods used by other crates/modules.

This change also replaces the boxed watch stream with a
`tokio_stream::wrappers::WatchStream`--also wrapped by a
profile-specific type--to avoid needless allocation and support
retaining additional resources.