retry: only wrap bodies when a request can be retried
Currently, the decision of whether or not a request can be retried is
made in the `clone_request` implementation for the `Retry` type in
`linkerd_app_core`. This is problematic, because by the time
`clone_request` is called, the request body has _already_ been wrapped
in a `ReplayBody`. The `ReplayBody` will buffer _regardless_ of whether
or not it's cloned. This means that the maximum length to buffer doesn't
_actually_ stop us from buffering --- it just determines whether we will
actually attempt to retry on failure, and whether we will retain that
buffer past when the request ends. The request will still buffer all
body data, but it will be dropped when the single instance of the body
is dropped. This means that for very long streaming bodies, we may
buffer increasingly large amounts of data as long as that request is in
flight.

This branch fixes this by moving the determination of whether a request
is retryable to *before* we wrap it in a `ReplayBody`, so that requests
over the maximum buffered length are not buffered. We do this by
introducing a new trait that serves as a predicate for whether or not a
request is retryable, and checking that predicate before wrapping the
body _or_ calling `clone_request`.

An alternative implementation which avoids adding the new trait would be
to change the `ReplayBody` type to only buffer data if the `Arc` around
the shared buffer state has more than one clone. That way, we could
still make the determination in `clone_request`. If we don't clone a
body, the body would know not to buffer. This would make the change
entirely internal to the `ReplayBody` type, rather than threading a new
trait through from the `Policy` type. I didn't go with this approach
because it felt kind of like spooky action at a distance, with the
body's behavior silently changing based on whether or not it's cloned,
but it would let us leave the rest of the code outside the
`linkerd_http_retry::replay` module as it is currently. If this approach
would be preferred, I'm happy to make that change instead.

In a follow-up, I'll also modify the replay body to accept a limit on
how much data it will buffer, and error if that limit is reached.
Although the change in this PR fixes the bug, it seems good to have some
defence-in-depth against unbounded buffering.