retry: only wrap bodies when a request can be retried (#1039)
Currently, the decision of whether or not a request can be retried is
made in the `clone_request` implementation for the `Retry` type in
`linkerd_app_core`. This is problematic, because by the time
`clone_request` is called, the request body has _already_ been wrapped
in a `ReplayBody`. The `ReplayBody` will buffer _regardless_ of whether
or not it's cloned. This means that the maximum length to buffer doesn't
_actually_ stop us from buffering --- it just determines whether we will
actually attempt to retry on failure, and whether we will retain that
buffer past when the request ends. The request will still buffer all
body data, but it will be dropped when the single instance of the body
is dropped. This means that for very long streaming bodies, we may
buffer increasingly large amounts of data as long as that request is in
flight.

This branch fixes this by moving the determination of whether a request
is retryable to *before* we wrap it in a `ReplayBody`, so that requests
over the maximum buffered length are not buffered. We do this by
introducing a new trait that serves as a predicate for whether or not a
request is retryable, and checking that predicate before wrapping the
body _or_ calling `clone_request`.

Co-authored-by: Oliver Gould <ver@buoyant.io>