build(deps): Skip coverage by PR title
e0e2b3407 updated the coverage action to be skipped when the
`dependencies` label is set, but in practice dependabot sets labels
after the PR is actually created, so the action can fire before the
label is present.

This change updates the coverage action to look at the PR title and skip
coverage when the title begins with `build(deps): `.