Remove unneeded uses of indexmap (#1048)
We initially adopted `indexmap` in many places where it's not strictly
necessary: we don't actually rely on the data being insert-ordered and
don't need to mutate the structure while iterating over it. In fact, the
traffic split module is the only place in the codebase where we access
the data structure by index (rather than by key).

In all other cases, we do not care about insertion-order. In the vast
majority of cases we can simply use a `HashMap`. Endpoint and route
labels need to be reliably ordered so that they can be compared (i.e.
implementing `Eq`), in which case the standard `BTreeMap` works
sufficiently.

This change eliminates unnecessary uses of `indexmap` to reduce
dependency and conceptual overhead.

Co-authored-by: Eliza Weisman <eliza@buoyant.io>