Replace the `request_error_total` metric (#1133)
We currently expose a single `request_errors_total` metric with a
`direction` label that indicates whether the error is on the inbound or
outbound proxy. This is in-line with many of our other metrics; but this
conflicts with prometheus best practices as we do not necessarily
maintain a uniform set of labels for all instances of a metric.  To
correct this, we want to adopt a scheme of using distinct metrics for
differing scopes with differing labels.

This change replaces the single `request_errors_total` metric with two
new metrics: `inbound_http_errors_total` and
`outbound_http_errors_total`.

As it turns out, the `request_errors_total` metric isn't actually
consumed anywhere in the _linkerd2_ repo, so this change can be made
safely without requiring control-plane changes. In follow-up changes, we
should probably include additional labels on these metrics to reflect
the stack targets; and then we should surface these errors elsewhere in
Linkerd.

This change also sets up introducing additional error metrics, for
instance to surface TLS detection errors, by decoupling the error-metrics
middleware from a specific metric.