build(deps): update `tonic`, `prost`, and  `linkerd2-proxy-api` (#1137)
This updates `prost`, `prost-types`, and `prost-build` to v0.8, which
includes a fix for a panic (and potential denial-of-service attack) when
converting a protobuf duration into a Rust `Duration`. Although we don't
use the vulnerable APIs in the proxy or in `linkerd2-proxy-api`, this is
necessary in order to fix a RUSTSEC advisory warning. In order to update
`prost`, we must also update `tonic` and `tonic-build` to v0.5, which
depends on `prost` 0.8, and update the `linkerd2-proxy-api` crate to
include linkerd/linkerd2-proxy-api#71.

Since these crates all depend on each other, we need to update them all
at the same time. Dependabot has opened separate PRs for these crates,
but none of them will pass CI, since they depend on incompatible
versions. This PR, on the other hand, should pass, since it updates all
the crates atomically in one commit. Also, some minor code changes
were required due to breaking API changes in `tonic` 0.5.

Closes #1134, #1135, and #1136; should fix CI.