Handle profile endpoint in Gateway outbound stack
In linkerd2/PR#6090, we noticed that when testing the StatefulSet
changes with a service that sends a request from the source cluster and
expects to be connected directly to a pod in the target cluster, the
outbound side of the Gateway in the target cluster throws a 'BadDomain'
error. When establishing a direct connection to a pod, the profile
returned from the look-up will contain an endpoint and it will not
contain a logical address.

The 'BadDomain' error stems from how the Gateway handles outbound
connections. If the profile does not include a logical address (i.e an
authorityOverride), then the connection is failed. This change adds
functionality to the Gateway to also consider endpoints received in the
profile on the outbound side. Support for endpoint handling on the
outbound side will unblock PR#6090.

To test, I have added a unit test that will check the endpoint stack is
considered when dealing with an endpoint in the profile. There are also
manual tests done in conjunction with the changes from PR#6090.

Signed-off-by: Matei David <matei@buoyant.io>