metrics: Fix metrics test code allowing incomplete matches (#1146)
Currently, there's a bug in the metrics test code where a metric can
match if it contains any of the expected labels, even if some are
missing. This is because the test code assumes a metric will match, and
sets the matched value to false if any expected label keys have
unexpected values. However, if a label that the matcher is configured to
expect isn't present at all in the actual metric, it is still
considered matched. This allows tests to pass even if required labels
aren't present, which is pretty bad.

This branch fixes that bug by counting how many matches were found, and
comparing it to the number of labels that are expected. If fewer labels
were found, then the metric is not a match, even if all the found labels
had the expected values.

It turns out that a lot of the existing tests expect labels that are not
actually present, and these were passing due to the bug. In particular,
we expected transport metrics to have authority labels, but we don't
actually add these to transport metrics, even when they were recorded
from HTTP connections for which we do know an authority. AFAICT, we
never added these labels, and it would be a pretty large change to add
them when possible, so I'm pretty sure the tests are just incorrect,
here. I've changed them not to expect these labels. But, if we actually
meant to be adding authority labels to these metrics all along, I can
try to fix that instead.

Similarly, we expected the inbound target_addr label to be added to
both peer="src" and peer="dst" transport metrics, but we don't
actually add it in the proxy. Again, I think this is just a bug in the test
code, so I've changed the tests to not expect it.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>