tests: remove `inbound_fuzz_addr`
Previously, this was used to replace the mocked SO_ORIGINAL_DST's IP
address with a totally bogus one. We could do this without breaking
stuff because the inbound proxy would still always forward on localhost.

It doesn't do that anymore, so the IP address part of the mock
SO_ORIGINAL_DST is actually load-bearing now. This commit removes
`inbound_fuzz_addr` from the integration tests.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>