Make ClientHandle internals private
There's no need to expose the inner workings of ClientHandle/Close (i.e.
that it's backed by an `Arc<Notify>`). Instead, we expose a
`ClientHandle::new` that builds a `ClientHandle` and a future that
completes when close has been called.

We test that the close future is called (but it isn't for some reason!).