fuzz: fix inbound fuzz logic expecting wrong IP address (#1193)
PR #1179 changed the inbound HTTP fuzz logic, and added a mocked
`Target` type similar to other stack tests. However, this introduced a
bug, because the `Target::addr()` function, which returns the socket
address that the mock TCP connector should expect the proxy to try to
connect to, returns the (mocked) original destination address. Since the
stack we build for the fuzz test maps the target IP address to localhost
before calling into the connect stack (as the real proxy currently
does), this means that we attempt to connect to an address the mock
connector doesn't expect. Then, it panics and the fuzz test fails.

This branch fixes this issue by changing `Target::addr` to return
127.0.0.1:80 instead of 192.0.2.2:80. This should match what the proxy
will actually attempt to connect to.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>