fuzzing: change `linkerd_dns::fuzz_target_1` to single-threaded (#1191)
* fuzzing: change `linkerd_dns::fuzz_target_1` to single-threaded

Something about the behavior of the fuzz logic for `linkerd_dns` is
hitting [a recently added assertion][1] in Tokio's multithreaded runtime
when shutting down. It's unclear what exactly triggers the assertion,
but it appears to be environment-dependent in some way: I can't
reproduce the failure on my machine using the reproducer generated by
`cargo-fuzz`, but it fails reliably in cluster-fuzz (possibly related to
the number of CPU cores on the cluster-fuzz servers?).

Since the purpose of this fuzz test is to fuzz the DNS name parsing
behavior, and this failure isn't actually related to the code being
fuzzed, just the environment the fuzz test runs in, this branch changes
the fuzz logic to use a single-threaded Tokio runtime. We'll continue
trying to get a minimal repro of the crash to report upstream, but this
should fix the fuzz test so that it can continue exercizing the DNS code
we actually care about here.

[1]: https://github.com/tokio-rs/tokio/blob/362df5a3172f6e1bdee2fd3808e5cfc730a111f6/tokio/src/runtime/thread_pool/worker.rs#L603

* use single-threaded rt in all fuzz targets

(or, in all the ones that use tokio, anyway)

Signed-off-by: Eliza Weisman <eliza@buoyant.io>