app: Defer stack construction until the proxy starts
We currently eagerly build the inbound and outbound stacks. In an
upcoming change, however, we'll want to ensure that we've obtained
inbound port policies before initializing the inbound stack.

This change modifies app initialization to move stack initialization
into the proxy's start task. In order to do this, we eagerly bind the
inbound/outbound listeners and provide the stream of incoming sockets to
the stacks at construction-time.

This doesn't actually change any behavior since stack construction was
moved into a task that was run lazily. The change is purely internal,
restructuring how/when this task is built.