app: Block proxy initialization on identity readiness (#1202)
Currently, the proxy may start serving connections even before identity
is provisioned. This won't be feasible once we introduce policy
discovery--the proxy will have to establish its inbound policy before
serving inbound connections.

This change modifies proxy initialization to block proxy initialization
on identity. When identity isn't established, warnings are logged every
15s.

While this is a behavior change from the proxy's point of view, this
shouldn't be a user-facing change, as container initialization is
blocked on the proxy's readiness, which is dependent on identity
initialization.