inbound: Use policy protocol configurations
Per-port server policies include protocol hinting that is not currently
honored.

This change modifies the inbound detect stacks to only perform HTTP
detection when the server policy indicates that detection should be
pefromed. It also honors the per-port detect timeout configuration. When
a port is documented as HTTP/1, HTTP/2, or gRPC, protocol detection is
skipped and the configuration is used instead.

When the port's protocol is documented to be application TLS, we check
that authorization policies apply but we do not explicitly enforce that
application TLS was present, since we may _additionally_ apply mesh TLS
in these situations.