policy: Add support for cluster-scoped default policies
The policy controller supports the default policies
`cluster-authenticated` and `cluster-unauthenticated` but the proxy
would error if configured with these policies. In order to use a single
policy annotation honored by both the policy controller and proxy, this
change adds support for a new configuration,
`LINKERD2_POLICY_CLUSTER_NETWORKS`, that allows the proxy to enforce
cluster-scoped defaults. When this configuration is not specified, the
cluster-scoped defaults are not supported.

This default will apply to all ports that are not documented in the
pod's spec (i.e. ports not in the `LINKERD2_PROXY_INBOUND_PORTS`
configuration).