inbound: Return HTTP-level authorization errors
The initial policy/authorization implementation only operates on
connections. When an connection is not authorized, the connection is
dropped and no information is passed back to the client. This also means
that all authorizations are made at connect-time and policy
changes--especially those that revoke access--are not observed
until a new connection is established. This is at the very least
awkward when trying to test policies.

This change modifies the inbound proxy's behavior for HTTP connections:
when a server configures a port to use (or be detected) as HTTP, the
authorization decision is now deferred until a request is processed. As
each request is handled, the policy's state is checked to determine
whether the connection is still permitted. If the connection is not
authorized, a `403 Forbidden` status code is returned for HTTP
requests (and the grpc-status `PermissionDenied` is used for gRPC
requests). The http error metrics also now reflect an `unauthorized`
reason.

This change, unfortunately, adds some wrinkles to our transport metric
labeling: HTTP connections will no longer have a `saz_name`
annotation--as no authorization policy is associated with these
connections (though they will continue to include the `srv_name` from
when the connection was established).