profiles: Avoid creating a default route stack
When the proxy resolves a service profile with routes that do not match
any requests, we handle these requests on a 'default' route. This route
does not have any labels and is generally superflous, inflating our
metrics export unnecessarily.

This change removes this behavior by instead skipping the route stack
entirely when no routes are matched. In order to do this, we add some
extra boxes so that we can handle the request & response types
transparently.